Move schedule creation to add page, match existing patterns
