+++ web-v4/src/routes/vehicles/[id]/schedules/+page.server.ts
@@ -46,44 +46,6 @@ export const actions: Actions = {
 		}
 	},
 
-	create: async ({ locals, params, request }) => {
-		const data = await request.formData();
-
-		let classificationId = Number(data.get('classification_id'));
-		const newName = data.get('new_name') as string | null;
-
-		if (newName) {
-			const keywords = (data.get('new_keywords') as string || '')
-				.split(',')
-				.map((k) => k.trim())
-				.filter(Boolean);
-			const classification = await createClassification(
-				params.id!,
-				{ classification: { name: newName, keywords } },
-				locals,
-			);
-			classificationId = classification.id;
-		}
-
-		const body: Record<string, unknown> = {
-			service_schedule: {
-				classification_id: classificationId,
-				distance_interval: data.get('distance_interval')
-					? Number(data.get('distance_interval'))
-					: null,
-				month_interval: data.get('month_interval') ? Number(data.get('month_interval')) : null,
-				notes: data.get('notes') || null,
-			},
-		};
-
-		try {
-			await createServiceSchedule(params.id!, body, locals);
-			return { success: true };
-		} catch {
-			return fail(422, { error: 'Failed to create schedule' });
-		}
-	},
-
 	suggest: async ({ locals, params, request }) => {
 		const data = await request.formData();
 		const presetNames: string[] = JSON.parse((data.get('preset_names') as string) || '[]');
